Adding class user and fix some issues
