+++ src/Users/Professor.java
@@ -1,81 +1,45 @@
-package Users;
+package users;
 
 import java.sql.Date;
 
-import observer.Observer;
-
-public class Professor implements Observer {
-	String name;
+public class Professor extends User {
 	String department;
 	Date hirringDate;
-	String PhDTopic;
-	String email;
-	String phoneNumber;
-	
-	
+	String phDTopic;
+
 	public Professor(String name, String department, Date hirringDate, String phDTopic, String email,
 			String phoneNumber) {
 		super();
 		this.name = name;
 		this.department = department;
 		this.hirringDate = hirringDate;
-		PhDTopic = phDTopic;
+		this.phDTopic = phDTopic;
 		this.email = email;
 		this.phoneNumber = phoneNumber;
 	}
-	
-	
-	public String getEmail() {
-		return email;
-	}
 
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
 	public String getDepartment() {
 		return department;
 	}
+
 	public void setDepartment(String department) {
 		this.department = department;
 	}
+
 	public Date getHirringDate() {
 		return hirringDate;
 	}
+
 	public void setHirringDate(Date hirringDate) {
 		this.hirringDate = hirringDate;
 	}
+
 	public String getPhDTopic() {
-		return PhDTopic;
+		return phDTopic;
 	}
+
 	public void setPhDTopic(String phDTopic) {
-		PhDTopic = phDTopic;
+		this.phDTopic = phDTopic;
 	}
-	
-	
-	public void notify(String message) {
-		// do some stuff
-		
-	}
-	
-	
-	
+
 }
