+++ src/Users/Student.java
@@ -1,62 +1,22 @@
-package Users;
+package users;
 
-import observer.Observer;
+public class Student extends User {
+	int id;
 
-public class Student implements Observer{
-	String name;
-	int Id;
-	String email;
-	String phoneNumber;
-	
-	
-	
 	public Student(String name, int id, String email, String phoneNumber) {
 		super();
 		this.name = name;
-		Id = id;
+		this.id = id;
 		this.email = email;
 		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
 	}
 
 	public int getId() {
-		return Id;
+		return id;
 	}
 
 	public void setId(int id) {
-		Id = id;
+		this.id = id;
 	}
-	
-	
-	public void notify(String message) {
-		// do some stuff
-	}
-	
+
 }
