+++ src/Users/TA.java
@@ -1,14 +1,7 @@
-package Users;
+package users;
 
-import observer.Observer;
-
-public class TA implements Observer{
-	String name;
+public class TA extends User {
 	String department;
-	String email;
-	String phoneNumber;
-	
-	
 
 	public TA(String name, String department, String email, String phoneNumber) {
 		super();
@@ -18,34 +11,6 @@ public class TA implements Observer{
 		this.phoneNumber = phoneNumber;
 	}
 
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
 	public String getDepartment() {
 		return department;
 	}
@@ -53,10 +18,5 @@ public class TA implements Observer{
 	public void setDepartment(String department) {
 		this.department = department;
 	}
-	
-	public void notify(String message) {
-		// do some stuff
-		
-	}
-	
+
 }
