+++ src/Users/User.java
@@ -1,0 +1,40 @@
+package users;
+
+import observer.Observer;
+
+public class User implements Observer {
+    String name;
+    String email;
+    String phoneNumber;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public void notify(String message) {
+        // do some stuff
+
+    }
+
+}
